[Parcelize] Split out PSI-only quick fix base class.
The bulk of the existing Parcelize quick fixes only rely on PSI, not on
analysis (the exception being ParcelMigrateToParcelizeQuickFix). To
enable running these quick fixes on K2, create a base class derived from
KotlinPsiOnlyQuickFixAction, and change the PSI-only quick fixes to
derive from that base class.

In addition, change the shortenReferences extension function to use
ShortenReferencesFacility, so that it will work on K2.

^KTIJ-25177